system-ops/requirements.txt
# NetPilot System Operations - Dependências Python
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
jinja2==3.1.2
httpx==0.25.2
pyyaml==6.0.1
cryptography==41.0.7
psutil==5.9.6
redis==5.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
paramiko==3.4.0
dnspython==2.4.2
certifi==2023.11.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
watchdog==3.0.0